pyproject.toml
[project]
authors = [
  {name = "Cleverson Sacramento", email = "cleverson.sacramento@gmail.com"},
]
dependencies = [
  "langgraph (>=0.6.8,<0.7.0)",
  "langgraph-prebuilt (>=0.6.4,<0.7.0)",
  "langchain-openai (>=0.3.33,<0.4.0)",
  "langchain-core (>=0.3.76,<0.4.0)",
  "prettyprinter (>=0.18.0,<0.19.0)",
  "authlib (>=1.6.4,<2.0.0)",
  "python-dotenv (>=1.0.1,<2.0.0)",
]
description = ""
name = "langgraph-experiments"
package-mode = true
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[tool.poetry]
packages = [
  {include = "src", from = "."}
]

[project.scripts]
langgraph-experiments = "src.main:main"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
pylint = "^3.3.8"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]
